refactor(utils): 非推奨 shim errorMessages.ts を削除し errorClassification に統一
再エクスポート shim（唯一の消費者は自テスト）を削除。shim テストのユニークな
2 アサーション（技術情報非漏洩・context 機密マスキング）を
errorClassification.test.ts に移植して coverage を維持。ErrorType の並列定義は
解消済み（errorClassification.ts の 1 箇所のみ）を grep で確認。
